Change how mobs are stored for faster lookup
This isn't a necessary optimization, the game runs perfectly fast
without it, but it makes some code shorter and a little cleaner.
Level.update() works now, but there's a commented out section of code
that iterates through mobs.values() from something I tried first. The
commented out code should work, but currently doesn't due to a bug in
SquidLib. The not-commented code works fine; getAt() is like a List's
get() method in that it takes an index instead of a key, and gets a
value from the OrderedMap.